test: rename initialize helper to reloadWithProductionTiming and document it
Helper files are named after what they export. Expand the doc comment to say
why the navigator.webdriver spoof is justified (loading states cannot exist
under test timing), what phase it belongs to (arrange), and why it needs no
explicit cleanup (each test gets a fresh incognito context and page).
