+++ src/e2e/puppeteer/__tests__/gestures.ts
@@ -4,9 +4,9 @@ import newThoughtCommand from '../../../commands/newThought'
 import $ from '../helpers/$'
 import exportThoughts from '../helpers/exportThoughts'
 import gesture, { endGesture, startGesture } from '../helpers/gesture'
-import reloadWithProductionTiming from '../helpers/initialize'
 import keyboard from '../helpers/keyboard'
 import paste from '../helpers/paste'
+import reloadWithProductionTiming from '../helpers/reloadWithProductionTiming'
 import scrollTo from '../helpers/scrollTo'
 import waitForSelector from '../helpers/waitForSelector'
 import { page } from '../session'
